Rename db folder to match the Gin template directories
