+++ database/redis.go
@@ -1,4 +1,4 @@
-package db
+package database
 
 import (
 	"github.com/beego/beego/v2/server/web"
